NetMeter: Make partial test results appear with a warning if Iperf failed to complete
Previous behavior marked a test as failed, even if partial results were available.
This fix enables to use NetMeter with older Iperf versions, where tests with small
packet sizes and multiple streams failed. It is however, STRONGLY recommended to
use NetMeter with the new Iperf versions only. The reason for this patch is a more
correct behavior, and not an attempt to make it usable with older Iperf versions.

This patch leaves a substantial code redundancy, which will be addressed to in a
later patch.

Signed-off-by: Leonid Bloch <leonid@daynix.com>