NetMeter: Check formatting of ssh credentials file
Now the creds file for ssh connection is checked for formatting errors,
and if such errors are encountered, an error message is displayed,
followed by the termination of the run.

This also has a security element to it: if the creds file was accidently
left unchanged after performing tests on a Windows guest, previously, an ssh
key file name was suggested that was the same as the password to the guest
Windows machine from before.

Signed-off-by: Leonid Bloch <leonid@daynix.com>