NetMeter: Change the logo encoding
Previously, the encoded logo contained its color profile, which is not
needed for its use - branding the html output. Removing the color
profile made the code for the logo two times shorter.

Additionally, the logo was slightly modified - the thin line surrounding
it, which is not clearly visible anyway, was removed. This made the text
look slightly bigger and clearer.

Also, the "daynix_logo" variable was swapped with just "logo" for
generalization and shortness.

Signed-off-by: Leonid Bloch <leonid@daynix.com>