NetMeter: Print notifications of tests that completely failed on the summary plots
Previously, for tests that failed completeley (not a single successful
attempt to communicate) nothing was plotted on the summary plot - these
points were just omitted.

Now, a label is printed there (when comparing results - labels of
different color are printed for each dataset) and the CPU utilization is
plotted at these points as well.

Signed-off-by: Leonid Bloch <leonid@daynix.com>