NetMeter: Bugfix: Last output per stream is now what it claims to be
Previously, the last output per stream in NetMeter was actually Iperf's
summary of that stream.
Now these summaries are ignored at parsing. It should be noted, that
since these summaries are not always accurate, and sometimes are missing
entirely from the current Iperf's output, they should not be used for
any purpose.

Following this fix, each run will take 2 s longer than previously, and
a small fix was needed for the plots' scaling, due to the slightly
different timing.

Signed-off-by: Leonid Bloch <leonid@daynix.com>