NetMeter: Considering a case when more streams than expected arrive
When a wrong number of streams arrive to the server, an error occurs.
Previously, an error message was correctly displayed only for the case
when less streams than expected arrived. Now, following encountering
such cases in tests, a correct error message will be displayed also when
more streams than expected arrive to the server.

Signed-off-by: Leonid Bloch <leonid@daynix.com>