NetMeter: Added ability to perform tests between any two machines
This patch brings a major broadening to the purpose of NetMeter.
Previously, NM was intended exclusively to measure network performance
between a virtual machine (VM) and its host, while running on the host
itself. Now, the performance can be measured between any two clients,
physical or virtual. Thus NetMeter becomes a general usage network
performance testing automation tool. The old functionality is preserved,
and can be used by setting the host as one of the clients and the VM as
the other.

The CPU usage is monitored now only if the machine that NetMeter runs on
is one of the clients.

Additionally, the testing can be performed between network controllers
that are different from those that NM is connected to.

If you have used NetMeter before, please familiarize yourself with the
new parameters in the NetMeterConfig.py file, and their usage.

NM_compare, and README are updated accordingly, to reflect these
changes. NM_compare will not work (without manual intervention) for
comparing results that were generated by previous NetMeter versions,
because of a different naming convention.

Signed-off-by: Leonid Bloch <leonid@daynix.com>