Fix deprecated NumPy expression
In the latest versions of NumPy creating an ndarray from ragged nested
sequences of different length is deprecated. However, it is still
possible to do if specifying 'dtype=object' when creating the ndarray.
But an 'object' ndarray can not be flattened, so using 'hstack' instead,
and converting the flattened array back to integer type, because
'lines_to_remove' needs to be of integer type to be used as an array
of indexes.

Signed-off-by: Leonid Bloch <leonidb@asocscloud.com>