Avoid comparing a string to NumPy numeric types
In combinations of recent Python and NumPy, comparing a string to NumPy
numeric type results in a warning:

```
FutureWarning: elementwise comparison failed; returning scalar instead,
but in the future will perform elementwise comparison
```
Avoid this, by verifying that the variable is a string before
comparison. This comparison is useful only when this variable is a
string anyway - for the summary plot.

Signed-off-by: Leonid Bloch <leonidb@asocscloud.com>